scripts/requirements.txt
python-telegram-bot==20.7
schedule==1.2.0
redis==5.0.1
aiohttp==3.9.1
aiohttp-cors==0.7.0